requirements.txt
asgiref==3.3.1
Django==3.0.14
psycopg2-binary==2.8.6
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
redis==4.3.4
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
Pillow==9.2.0
requests==2.27.1
haversine==2.5.1
kombu==5.3.2
celery==5.3.4
drf-yasg==1.21.7
boto3==1.28.68